requirements.txt
PyJWT==1.7.1
Flask-SQLAlchemy==2.4.1
Flask==1.1.2
Werkzeug==1.0.1